+++ AnS/UI/Mixins/Settings/General.lua
@@ -18,9 +18,7 @@ local tooltipRealm = {
 };
 
 local tooltipRegion = {
-    "tooltipRegionRecent",
     "tooltipRegionMin",
-    "tooltipRegion3Day",
     "tooltipRegionMarket"
 };