+++ AnsAuctionData/Data.lua
@@ -1,5 +1,3 @@
 local Ans = select(2, ...);
-local Realms = {};
-local Connected = {};
-Ans.Realms = Realms;
-wipe(Connected);
+local R = {};
+Ans.Realms = R;